+++ src/Page/Admin/Store/AdminStoreCreate/index.ts
@@ -1,0 +1,1 @@
+export { default as AdminStoreCreate } from "./AdminStoreCreate";
